+++ ExpensesManager.h
@@ -21,7 +21,7 @@ public:
     void showAllExpenses( void );
     vector <Expense> selectExpensesBetweenBateBorders( int , int );
     void showSortedExpensesBetweenDateBorders( int , int );
-    int sumExpensesAmountsBetweenDateBorders( int , int );
+    float sumExpensesAmountsBetweenDateBorders( int , int );
     vector <Expense> sortExpensesVector( vector <Expense> );
     void showChosenExpenses( vector<Expense> * );
 };
